livekit-escalation/pyproject.toml
[project]
name = "livekit-escalation"
version = "1.0.0"
description = "LiveKit video escalation for CaritaHub Customer Service Kiosk"
requires-python = ">=3.11,<3.15"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "livekit-api>=0.7.0",
    "clickhouse-connect>=0.8.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]

[project.scripts]
escalation-server = "orchestrator.server:main"

[tool.setuptools.packages.find]
include = ["orchestrator*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
